rm input data
Was reading the about page and noted that the author of AoC doesn't like
inputs to be commited if possible. I was under the impression that the
inputs were specific to each user. But for now, until I add in file
reading I suppose I'll just quickly erase this in good faith