Day 9: lunch break fiddling
I think I've got the fitting logic down
but I see my bug, not quite iterating leftward from the end in the same
way that the example does. They only ever say "This thing on the right,
where does it fit?" and they do it in that order and never again.

Which is why there's a huge hole in their example where 8888 would fit
but it does't move