Day 19 Part 1 first attempt
Is it just taking a while or is it infinite looping? I'm not sure.
It worked on the sample input so it seems like the idea to convert each
provided pattern into a small parsing engine, then check the input
against each parser, consuming the input matched so far then checking
each potentialy parser against that worked in a small scale.

But I suppose this is going to be a bit slow on a larger input. I could
probably use the RegEx library and it would be significantly faster, but
that would be passing up a chance to write a little parser which is no
fun.