Day 19 part 2 pre-work checkpoint
I have an idea that's mostly working, it fails on one of the inputs and
I haven't had the chance to figure out why yet, but I will.

I'm pretty sure that I could use a Trie for this because that does
prefixing search for stuff like autocomplete and this is sort of similar
as far as problems go, but I haven't looked too far into this idea yet.