Day 19 Part 2: port python solution from earlier to rust
I took some time during a break at work to get a Trie in python that
worked out fine, figured it would take some time to make in rust and so
waited a bit but submitted the answer earlier. This was easier than I
thought besides the fact that I think I _was_ definitely going in the
right direction with my previous code, I just had the data structure not
setup so correctly.

For the time being the Trie is indexed by char rather than one of my
towel stripes because I didn't want to think about how I might need to
modify the way the slices work and such to look for a length of the
enum. Though I think maybe it would have worked anyway though but eh...
it works and it works so that's good for now.